+++ jungle-examples/examples/welcome/animals/rhythm_guitarist.rs
@@ -41,16 +41,16 @@ impl Animal for RhythmGuitarist {
 }
 
 const RHYTHM_GUITAR_LANE_ID: u32 = <<RhythmGuitarist as Animal>::Id as AnimalIdValue>::U32;
-type Pick<const NOTE: u8, const NOTE_TICK: u8, const REST_TICK: u8> =
+type Pick<const NOTE: u8, const NOTE_TICK: u32, const REST_TICK: u32> =
     LanePick<NOTE, NOTE_TICK, REST_TICK, RHYTHM_GUITAR_LANE_ID>;
-type Pluck<const NOTE_1: u8, const NOTE_2: u8, const NOTE_TICK: u8, const REST_TICK: u8> =
+type Pluck<const NOTE_1: u8, const NOTE_2: u8, const NOTE_TICK: u32, const REST_TICK: u32> =
     LanePluck<NOTE_1, NOTE_2, NOTE_TICK, REST_TICK, RHYTHM_GUITAR_LANE_ID>;
 type Strum<
     const NOTE_1: u8,
     const NOTE_2: u8,
     const NOTE_3: u8,
-    const NOTE_TICK: u8,
-    const REST_TICK: u8,
+    const NOTE_TICK: u32,
+    const REST_TICK: u32,
 > = LaneStrum<NOTE_1, NOTE_2, NOTE_3, NOTE_TICK, REST_TICK, RHYTHM_GUITAR_LANE_ID>;
 
 pub struct IntroSectionMeta;
