+++ jungle-examples/examples/welcome/instrumentation/hihat/mod.rs
@@ -39,9 +39,9 @@ impl Instrument for HiHat {
     }
 }
 
-pub struct Chick<const NOTE: u8, const NOTE_TICK: u8, const REST_TICK: u8, const LANE_ID: u32 = 0>;
+pub struct Chick<const NOTE: u8, const NOTE_TICK: u32, const REST_TICK: u32, const LANE_ID: u32 = 0>;
 #[jungle::act]
-impl<const NOTE: u8, const NOTE_TICK: u8, const REST_TICK: u8, const LANE_ID: u32> Act
+impl<const NOTE: u8, const NOTE_TICK: u32, const REST_TICK: u32, const LANE_ID: u32> Act
     for Chick<NOTE, NOTE_TICK, REST_TICK, LANE_ID>
 {
     type Effect = Monad<HiHat, HiHatArticulation, LANE_ID, NOTE, NOTE_TICK, REST_TICK>;
