+++ jungle-examples/examples/welcome/instrumentation/snare_drum/mod.rs
@@ -40,9 +40,9 @@ impl Instrument for SnareDrum {
     }
 }
 
-pub struct Crack<const NOTE: u8, const NOTE_TICK: u8, const REST_TICK: u8, const LANE_ID: u32 = 0>;
+pub struct Crack<const NOTE: u8, const NOTE_TICK: u32, const REST_TICK: u32, const LANE_ID: u32 = 0>;
 #[jungle::act]
-impl<const NOTE: u8, const NOTE_TICK: u8, const REST_TICK: u8, const LANE_ID: u32> Act
+impl<const NOTE: u8, const NOTE_TICK: u32, const REST_TICK: u32, const LANE_ID: u32> Act
     for Crack<NOTE, NOTE_TICK, REST_TICK, LANE_ID>
 {
     type Effect = Monad<SnareDrum, SnareDrumArticulation, LANE_ID, NOTE, NOTE_TICK, REST_TICK>;
